Created Country select modal to reuse and combined the create trip modal into one file
